[leetCode] MedianTwoSortedArrays solution
This is my solution for the median two sorted arrays leetcode problem.

Was kinda tough, but simple in its own way. Tooka bout an hour to come
up with a solution. There are a couple of points of improvement, but
overall, it satisfies my own restrictions.

It must be readable and easy to understand, and it must pass all test
cases that leetcode provides.

Signed-off-by: Nathan <nihongotraveler@icloud.com>